requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.12.01
anthropic>=0.40.0
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
tiktok-uploader>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pysrt>=1.1.2
